@@ -1,14 +1,17 @@
 package org.phenoapps.verify;
 
-import android.content.SharedPreferences;
 import android.os.Bundle;
-import android.preference.Preference;
-import android.support.v7.app.AppCompatActivity;
+
+import androidx.appcompat.app.AppCompatActivity;
 import android.view.MenuItem;
 
 public class SettingsActivity extends AppCompatActivity {
 
     public static String FILE_NAME = "org.phenoapps.verify.FILE_NAME";
+    public static String INTRO_BUTTON = "org.phenoapps.verify.INTRO";
+
+    public static String ABOUT_BUTTON = "org.phenoapps.verify.ABOUT";
+
     public static String SCAN_MODE_LIST = "org.phenoapps.verify.SCAN_MODE";
     public static String AUDIO_ENABLED = "org.phenoapps.verify.AUDIO_ENABLED";
     public static String TUTORIAL_MODE = "org.phenoapps.verify.TUTORIAL_MODE";
